fix: resolve three code-quality issues
1. configCache mtime invalidation (web/lib/docs/fs.ts)
   - Replace a module-level singleton with an mtime-keyed entry so
     .docstudio.json changes are picked up without process restart.
   - Surface JSON parse errors instead of silently falling back to {}.

2. resolveProjectRoot multi-project support (packages/core)
   - resolveProjectRoot now returns {repoRoot}/content/projects/{projectId}
     when a projectId is provided, enabling workspace-level project lookup.
   - Eliminate the double-resolution bug: createProjectPathContract now
     accepts an explicit projectRoot and no longer calls resolveProjectRoot
     internally. All callers (content-repository, workflow-sync-service,
     init-service, tests) pass the already-resolved root.
   - Update project-contract test to write fixtures under the new workspace
     layout so the projectId-mismatch assertion still fires correctly.

3. local-studio-app.tsx megacomponent split (packages/web)
   - Extract all 22 pure utility functions and 8 shared types/constants
     into local-studio-utils.ts (independently testable, no React deps).
   - Extract workflow state (9 useState, 3 useEffect, 4 callbacks, derived
     values) into useWorkflowState custom hook (use-workflow-state.ts).
   - Main component reduced from 2 443 → 1 872 lines (−23 %).
